pyproject.toml
[tool.poetry]
name = "recruitmentportal_project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.6"
Pillow = "^9.4.0"
django-crispy-forms = "^1.14.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.0.2"}
django-multiselectfield = "^0.1.12"
shortuuid = "^1.0.11"
django-import-export = "^3.0.2"
django-taggit = "^3.1.0"
django-filter = "^22.1"
django-autocomplete-light = "^3.9.4"
social-auth-app-django = "^5.0.0"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
